# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JOB API",
        "description": "API da plataforma JOB para intermediação de serviços.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/users-categorias/categorias": {
            "get": {
                "tags": [
                    "Usuário-Categorias"
                ],
                "summary": "Lista todas as categorias e informa quais pertencem ao usuário autenticado",
                "operationId": "8f1f2cefb3c0b94fb8d1af28fa49e4cc",
                "responses": {
                    "200": {
                        "description": "Lista de categorias"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users-categorias": {
            "get": {
                "tags": [
                    "Usuário-Categorias"
                ],
                "summary": "Lista as categorias do usuário autenticado",
                "operationId": "c97d51e623d16cab6dbcb45a9f9d2f5c",
                "responses": {
                    "200": {
                        "description": "Lista de vínculos"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuário-Categorias"
                ],
                "summary": "Vincula uma categoria ao usuário autenticado",
                "operationId": "5fcc1380ee0e7ffca14a7c8ff70a78e4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "categoria_id"
                                ],
                                "properties": {
                                    "categoria_id": {
                                        "type": "integer",
                                        "example": 4
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vínculo criado"
                    },
                    "409": {
                        "description": "Categoria já vinculada"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users-categorias/{usuarioCategoria}": {
            "delete": {
                "tags": [
                    "Usuário-Categorias"
                ],
                "summary": "Remove uma categoria do usuário autenticado",
                "operationId": "e0ff58df76b7ca0c899aeffb66d8e0b0",
                "parameters": [
                    {
                        "name": "usuarioCategoria",
                        "in": "path",
                        "description": "ID do vínculo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Vínculo removido"
                    },
                    "404": {
                        "description": "Vínculo não encontrado"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticação do Laravel Sanctum",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}